requirements.txt
asgiref==3.3.4
Django~=3.2.4
django-rest-api==0.1.5
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1

decouple~=0.0.7
stripe~=2.58.0